+++ components/news/subscribeNews/subscribeNews.js
@@ -2,6 +2,11 @@ import { subscribeData } from "../../../data/subscribeData.js";
 import { shortenStr } from "../../../util/shortenStr.js";
 import { getSubscribeList } from "../../..//util/getSubscribeList.js";
 
+/**
+ *
+ * @param {number} brandIdx
+ * brandIdx === -1이면 구독한 언론사가 없습니다 렌더링
+ */
 export const subscribeNews = (brandIdx) => {
   // 로컬스토리지에서 받아온 brandId를 사용하여 언론사 뉴스 추출
   let brandId = getSubscribeList()[brandIdx];
@@ -21,11 +26,20 @@ export const subscribeNews = (brandIdx) => {
   // sub news selector
   const subNewsList = pressNews.querySelector(".sub-news-list");
 
+  // 구독한 언론사가 없는 경우
+  if (brandIdx === -1) {
+    renderNoSubscribe(pressNews);
+    return;
+  }
+
   // news Id 저장
   pressNews.dataset.brandIdx = brandIdx;
 
-  // 구독한 언론사 화면이므로 cateId = -1로 저장
-  pressNews.dataset.cateId = -1;
+  // brandId 업데이트
+  pressNews.dataset.brandId = brandId;
+
+  // 구독한 언론사 화면이므로 cateIdx = -1로 저장
+  pressNews.dataset.cateIdx = -1;
 
   // 언론사 로고
   brandLogo.src = newsData.brandImg;
@@ -47,4 +61,14 @@ export const subscribeNews = (brandIdx) => {
     )}</a>`;
     subNewsList.innerHTML += listItem;
   });
+};
+
+/**
+ *
+ * 구독한 언론사가 없는 경우 구독 유도문구 렌더링
+ * @param pressNews
+ */
+const renderNoSubscribe = (pressNews) => {
+  pressNews.classList.add("hidden");
+  document.querySelector(".no-subscribe").classList.remove("hidden");
 };
