+++ components/subscribe/style.css
@@ -1,0 +1,12 @@
+.subscribe-btn {
+  padding: 2px 8px;
+}
+
+.remove-subscribe-btn {
+  padding: 7px;
+}
+
+.no-subscribe {
+  padding: 100px;
+  font-size: 1.2rem;
+}
