+++ index.js
@@ -1,9 +1,9 @@
 import { subscribeSwitchBtn } from "./components/subscribeSwitchBtn/subscribeSwitchBtn.js";
 import { header } from "./components/header/header.js";
 import { addSubscribe } from "./components/subscribe/addSubscribe.js";
-import { removeSubscribtion } from "./components/subscribe/removeSubscribtion.js";
+import { removeSubscribe } from "./components/subscribe/removeSubscribe.js";
 
 header();
 subscribeSwitchBtn();
 addSubscribe();
-removeSubscribtion();
+removeSubscribe();
