+++ style.css
@@ -3,6 +3,7 @@
 @import url("./style/reset.css");
 @import url("./components/subscribeSwitchBtn/style.css");
 @import url("./components/fieldTab/style.css");
+@import url("./components/subscribe/style.css");
 
 .news-stand {
   width: 930px;
@@ -27,16 +28,6 @@
 
 .news-list {
   position: relative;
-
-  .press-info {
-    .subscribe-btn {
-      padding: 2px 8px;
-    }
-
-    .remove-subscribe-btn {
-      padding: 7px;
-    }
-  }
 
   .news {
     align-items: start;
